services/voice_watcher/requirements.txt
watchdog>=3.0.0
python-dotenv>=1.0.0
faster-whisper>=1.0.0
requests>=2.31.0
torch>=2.0.0

